fix compat with latest Verus
Recent Verus commit ea799f53e401e6f22e02965f446c7ab0916ab684 changed the
meaning of the generic parameter to `old()`, so the previous workaround
that relied on `old::<&mut _>()` failed to compile.  This commit
changes the workaround to `&*old()`, which was the other suggestion from
https://github.com/verus-lang/verus/issues/671 and is compatible with
both old and new Verus versions.