cleaner linearization proof for rwkv_inv_v
No need to linearize after commit in the case when the underlying
durable state didn't observably change (e.g., when doing a mutating
operation that doesn't actually change the state).  Instead, we can
prove a contradiction if we couldn't linearize at the disk write but
then succeeded in flushing and commit returned successfully.